% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract Function}
\usage{
extract(prefix, subj_nums, filepath)
}
\arguments{
\item{prefix}{A character prefix for the filename specifying what type of
file you want to pull the data from. Ex: "DF"}

\item{subj_nums}{A vector of the precise numbers of the subjects you are
interested in. Ex: 21:30}

\item{filepath}{A string giving the path to the data files you want
extracted}
}
\value{
Returns a list of dataframes, with each dataframe
belonging to a different subject.
}
\description{
This function extracts the data from the multiple files generated from the
experiment and stores it into a list of dataframe objects, one for each
subject.
}
\note{
Before calling this function, make sure you're working directory is in
the file you want to extract the data from.
}
\examples{
\dontrun{df <- extract("EB", 11, "./inst/extdata/")
df}

}
