% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plot.R
\name{generic_plot}
\alias{generic_plot}
\title{Generic Plot Function}
\usage{
generic_plot(data = x, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{timestamp}{A measure of time that will be converted to hours}

\item{x}{A data frame}
}
\value{
The function returns a visual of plotted data
}
\description{
This function plots two different scenarios. One of the scenarios is a
cumulative one. Essentially, the cumulative case takes the result of
calculations created by the extract, mpc, and var_def functions; it then
plots them. The second scenario has this function plotting specific session
wide variables.
}
\examples{
PUT SOMETHING HERE

}
