% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{med}
\alias{med}
\title{Helper functions that creates object of the med class}
\usage{
med(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
The function returns an object of class med.
}
\description{
This function takes a matrix and returns an object of class med.
}
