% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc}
\alias{mpc}
\title{MPC to R Data Frame}
\usage{
mpc(x, array, timescale = "min", cummulative = TRUE)
}
\arguments{
\item{x}{A list of mpc files for a single subject. These mpc files must first
be converted into .txt file types before being imported to R.}

\item{array}{A character string identifying the array to be extracted. This
character string should contain a single letter.}

\item{timescale}{A character string identifying the timescale that timestamps
should be converted to. Available timescales are: "hour", "min", "sec",
and "cent".}

\item{cummulative}{A logical indicating whether timestamps should be
merged cumulatively across sessions. If FALSE, time will be treated
independently for each session.}
}
\value{
A data frame containing subject id, session id, timestamps, and
associated event tags.
}
\description{
This function takes a sequential list of mpc files for an individual subject,
extracts a single, user-defined array,
and creates a single data frame containing
time stamps and event tags. Specifically, this function is designed to handle
arrays structured with a timestamp with a decimal that identifies event type.
}
\examples{
\dontrun{ x <- extract("EB", 11)
df <- mpc(x[[1]], "d")
df
}
}
