% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_med}
\alias{new_med}
\title{Helper functions that creates object of the med class}
\usage{
new_med(x)
}
\arguments{
\item{x}{The function returns a data frame.}
}
\value{
The function returns an object of class med.
}
\description{
This function takes the output of the 'var_def' function. It then makes sure
all of the data is in long form. After making sure the data is in the
right form, it then turns it into an object of class med.
}
