% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plot.R
\name{plot.med}
\alias{plot.med}
\title{Generic Plot Function}
\usage{
\method{plot}{med}(x, y, measure, time_stamps = time_stamps, value = value, ...)
}
\arguments{
\item{x}{A data frame that contains the values from our research.}

\item{y}{This variable is ignored.}

\item{measure}{This variable allows the user to specify what was being
measured in the trial.}

\item{time_stamps}{Variable that will be plotted on the x-axis}

\item{value}{Variable that will be plotted on the y-axis}

\item{...}{Placeholder for further arguments.}
}
\value{
The function returns a visual of plotted data.
}
\description{
This function takes in our our newly cleaned data. It then plots the data
allowing researchers to see a visual representation of their data.
}
