% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{select}
\alias{select}
\title{Selects a variable for plotting.}
\usage{
select(x, variable)
}
\arguments{
\item{x}{A data frame.}

\item{variable}{The variable of interest, will be the second column in the
matrix.}
}
\value{
The function returns an object of class med.
}
\description{
This function takes a data frame from the var_def function and selects
a variable of interest. A matrix is then generated and converted into the
med class that can then be passed into a generic plotting function to
generate a cummulative record.
}
