% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_def.R
\name{var_def}
\alias{var_def}
\title{Title}
\usage{
var_def(df, event_tags, cumulative = TRUE)
}
\arguments{
\item{df}{Describe parameter here}

\item{event_tags}{Describe parameter here}

\item{cumulative}{Describe parameter here}
}
\value{
Describe what it returns here
}
\description{
Function description
}
