% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plot.R
\name{generic_plot}
\alias{generic_plot}
\title{Generic Plot Function}
\usage{
generic_plot(x, y, measure, time_stamp, value, ...)
}
\arguments{
\item{x}{A data frame that contains the values from our research.}

\item{y}{This variable is ignored.}

\item{time_stamp}{Variable that will be plotted on the x-axis}

\item{value}{Variable that will be plotted on the y-axis}

\item{...}{Placeholder for further arguments.}

\item{timestamp}{A measure of time that will be converted to hours}
}
\value{
The function returns a visual of plotted data
}
\description{
ADD THE DESCRIPTION HERE
}
