% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plot.R
\name{plot.med}
\alias{plot.med}
\title{Generic Plot Function}
\usage{
\method{plot}{med}(x, y, ...)
}
\arguments{
\item{x}{A matrix that contains the values from our research.This is an
object of class med.}

\item{y}{This variable is ignored.}

\item{...}{Placeholder for further arguments.}
}
\value{
The function returns a visual of plotted data.
}
\description{
This function takes in our our newly cleaned data. It then plots the data
allowing researchers to see a visual representation of their data.
}
