% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_def.R
\name{var_def}
\alias{var_def}
\title{Define Dummy Variables}
\usage{
var_def(df, event_tags, cumulative = TRUE)
}
\arguments{
\item{df}{Dataframe that is being fed into the function. It is essential
that this dataframe be output by the mpc function, also defined in this
package. You may pass in a different dataframe but the format must be
identical to that of an mpc output.}

\item{event_tags}{A dictionary consisting of key value pairs. The
keys should correspond to all unique "event code" entries and the values
should be event names. These names will become the columns of the output
dataframe.}

\item{cumulative}{If set to TRUE, the function will also output columns
that keep track of the total number of times a particular event took place.}
}
\value{
This function returns a dataframe x, whose first three columns
are identical to the input dataframe, df, but with additional columns
including dummy variables for unique events and corresponding cumulative
columns.
}
\description{
Define Dummy Variables
}
