% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_def.R
\name{var_def}
\alias{var_def}
\title{Variable Define Function}
\usage{
var_def(df, event_tags, cumulative = TRUE)
}
\arguments{
\item{df}{Dataframe that is being fed into the function.}

\item{event_tags}{Describes which event took place.}

\item{cumulative}{Whether or not time is being added over sessions.}
}
\value{
This function returns
}
\description{
Function description
}
